Added start of a nn approach to lemmatizer
