+++ sentence_disambiguator/create_toy_data.py
@@ -123,36 +123,3 @@ def load_data(
                 else:
                     yield cur_doc_word
 
-
-def split_data(in_data, out_data=None, frac=0.80):
-    ''' splitting dataset into training/testing sets
-
-        Args
-        ----
-        in_data : np array
-                  input features x rows
-
-        out_data : np array
-                   output features x rows
-
-        frac : float
-               %  to keep in training set
-
-        Returns
-        -------
-        (training_inputs, testing_inputs),
-        [(training_outputs, testing_outputs)]
-    '''
-
-    num_data = in_data.shape[0]
-    num_tr_data = int(np.floor(num_data * frac))
-    indices = np.random.permutation(num_data)
-
-    tr_idx, te_idx = indices[:num_tr_data], indices[num_tr_data:]
-    tr_in_data, te_in_data = in_data[tr_idx, :], in_data[te_idx, :]
-
-    if not out_data is None:
-        tr_out_data, te_out_data = out_data[tr_idx, :], out_data[te_idx, :]
-        return (tr_in_data, te_in_data), (tr_out_data, te_out_data)
-
-    return (tr_in_data, te_in_data)
